Rename workers module to relays
