+++ relayer/relays/beefy/beefy-relaychain-listener.go
@@ -12,8 +12,8 @@ import (
 	"golang.org/x/sync/errgroup"
 
 	"github.com/snowfork/polkadot-ethereum/relayer/chain/relaychain"
+	"github.com/snowfork/polkadot-ethereum/relayer/relays/beefy/store"
 	"github.com/snowfork/polkadot-ethereum/relayer/substrate"
-	"github.com/snowfork/polkadot-ethereum/relayer/workers/beefy/store"
 )
 
 type BeefyRelaychainListener struct {
