+++ relayer/relays/beefy/store/store_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/sirupsen/logrus"
 	"github.com/snowfork/go-substrate-rpc-client/v3/types"
-	"github.com/snowfork/polkadot-ethereum/relayer/workers/beefy/store"
+	"github.com/snowfork/polkadot-ethereum/relayer/relays/beefy/store"
 	"github.com/stretchr/testify/suite"
 	"golang.org/x/sync/errgroup"
 )
